Rename file generalize module
